User search / lookup, mobile optimizations, renaming
