+++ apps/app/src/components/Social/FeedInsights/ScoreGraph.tsx
@@ -17,7 +17,7 @@ import { format, formatDistance } from 'date-fns'
 import throttle from 'lodash/throttle'
 import { useCallback, useMemo, useState } from 'react'
 import { useFeed } from '../../../contexts/feed'
-import { scoreEntry } from '../../../workers/scoring'
+import { scoreEntry } from '../../../workers/top/scoring'
 import { Entry } from '@riftdweb/types'
 import { PostTime } from '../_shared/PostTime'
