+++ apps/app/src/contexts/skynet/index.tsx
@@ -15,7 +15,7 @@ import { FeedDAC } from 'feed-dac-library'
 import { UserProfileDAC } from '@skynethub/userprofile-library'
 import { SocialDAC } from 'social-dac-library'
 import { ControlRef, useControlRef } from './ref'
-import { clearEntriesBuffer } from '../../workers/workerFeedLatest'
+import { clearEntriesBuffer } from '../../workers/feedAggregator'
 import { clearAllTaskQueues } from '../../shared/taskQueue'
 import { createLogger } from '../../shared/logger'
 import { clearAllTokens } from '../../workers/tokens'
