+++ apps/app/src/workers/top/scoring.ts
